+++ tools/tools-common.c
@@ -762,9 +762,13 @@ tools_print_events(const char *prefix, struct xkb_state *state,
                 // TODO
                 break;
             }
+            case XKB_EVENT_TYPE_POINTER_BUTTON: {
+                // TODO
+                break;
+            }
             default: {
-                static_assert(XKB_EVENT_TYPE_POINTER_MOTION == 5 &&
-                              XKB_EVENT_TYPE_POINTER_MOTION ==
+                static_assert(XKB_EVENT_TYPE_POINTER_BUTTON == 6 &&
+                              XKB_EVENT_TYPE_POINTER_BUTTON ==
                               (enum xkb_event_type) _LAST_XKB_EVENT_TYPE,
                               "Missing event type");
             }